+++ app/components/field_service/campaigns/form_page_component.html.erb
@@ -8,6 +8,9 @@
     <%= f.input :name %>
   <% end %>
   <%= input_wrapper do %>
+    <%= f.input :description %>
+  <% end %>
+  <%= input_wrapper do %>
     <%= f.input :start_date, include_blank: true %>
   <% end %>
   <%= input_wrapper do %>
