+++ app/components/field_service/campaigns/list_item_component.rb
@@ -5,8 +5,6 @@ class FieldService::Campaigns::ListItemComponent < ApplicationComponent
 
   has :campaign
 
-  private
-
   def actions
     [
       assignments_action,
@@ -14,26 +12,5 @@ class FieldService::Campaigns::ListItemComponent < ApplicationComponent
       show_action,
       delete_action
     ]
-  end
-
-  def assignments_action
-    link_to(t('app.links.assignments'), urls.field_service_campaign_assignments_path(campaign), class: 'btn')
-  end
-
-  def edit_action
-    link_to(t('app.links.edit'), urls.edit_field_service_campaign_path(campaign), class: 'btn')
-  end
-
-  def show_action
-    link_to(t('app.links.view'), urls.field_service_campaign_path(campaign), class: 'btn')
-  end
-
-  def delete_action
-    link_to(
-      t('app.links.delete'),
-      urls.field_service_campaign_path(campaign),
-      data: { method: :delete, confirm: t('app.messages.confirm_delete') },
-      class: 'btn'
-    )
   end
 end
