+++ app/components/field_service/campaigns/show_page_component.html.erb
@@ -1,1 +1,14 @@
-<div>Add FieldService::Campaigns::ShowPage template here</div>
+<%= render breadcrumb %>
+
+<%= render DropdownMenuComponent.new(classes: ['btn-primary'], type: :list_options) do |component| %>
+  <% actions.each do |action| %>
+    <% component.item do %>
+      <%= action %>
+    <% end %>
+  <% end %>
+<% end %>
+
+<div><%= render attribute(:name).with_label.without_icon %></div>
+<div><%= render attribute(:code).with_label.without_icon %></div>
+<div><%= render attribute(:description).with_label.without_icon %></div>
+<div><%= render attribute(:dates).with_label.without_icon %></div>
