+++ app/components/field_service/campaigns/show_page_component.rb
@@ -1,7 +1,26 @@
 # frozen_string_literal: true
 
 class FieldService::Campaigns::ShowPageComponent < PageComponent
-  def initialize(campaign:)
-    @campaign = campaign
+  include FieldServiceCampaignConcern
+
+  has :campaign
+
+  def actions
+    [
+      assignments_action,
+      edit_action,
+      delete_action
+    ]
+  end
+
+  private
+
+  def setup_breadcrumb
+    breadcrumb.add_item(
+      t('app.links.preaching_campaigns'),
+      urls.field_service_campaigns_path
+    )
+
+    breadcrumb.add_item(campaign.code)
   end
 end
