+++ app/components/page_component.rb
@@ -1,6 +1,14 @@
 # frozen_string_literal: true
 
 class PageComponent < ApplicationComponent
+  def initialize(*args)
+    super
+
+    if respond_to?(:setup_breadcrumb, true)
+      send(:setup_breadcrumb)
+    end
+  end
+
   def breadcrumb
     @breadcrumb ||= BreadcrumbComponent.new.tap do |b|
       b.add_item(t('app.links.home'), urls.root_path)
